Sicoob/requirements.txt
requests
python-dotenv
pocketbase
schedule